[language][move-cli] remove diem-dependencies
This removes all diem (crate) dependencies from the move-cli by
splitting its diem parts into a separate tool called df-cli.

Closes: #8615
