+++ Cargo.toml
@@ -67,6 +67,7 @@ members = [
     "language/compiler/ir-to-bytecode/syntax",
     "language/diem-framework",
     "language/diem-framework/releases",
+    "language/diem-tools/df-cli",
     "language/diem-tools/diem-events-fetcher",
     "language/diem-tools/diem-read-write-set",
     "language/diem-tools/diem-validator-interface",
@@ -197,6 +198,7 @@ default-members = [
     "language/tools/move-cli",
     "language/tools/move-coverage",
     "language/tools/move-unit-test",
+    "language/diem-tools/df-cli",
     "language/diem-tools/diem-events-fetcher",
     "language/diem-tools/transaction-replay",
     "language/diem-tools/writeset-transaction-generator",
