+++ language/diem-tools/transaction-replay/src/lib.rs
@@ -174,7 +174,12 @@ impl DiemDebugger {
             }
         }
         for event in o.events() {
-            state_view.save_contract_event(event.clone())?
+            state_view.save_event(
+                event.key().as_bytes(),
+                event.sequence_number(),
+                event.type_tag().clone(),
+                event.event_data().to_vec(),
+            )?
         }
         Ok(())
     }
