+++ language/move-stdlib/src/tests.rs
@@ -14,3 +14,18 @@ fn check_that_docs_are_updated() {
         "Generated docs differ from the ones checked in"
     )
 }
+
+#[test]
+fn check_that_the_errmap_is_updated() {
+    let temp_file = tempfile::NamedTempFile::new().unwrap();
+
+    crate::build_error_code_map(&temp_file.path().to_string_lossy());
+
+    assert!(
+        file_diff::diff(
+            &temp_file.path().to_string_lossy(),
+            &crate::move_stdlib_errmap_full_path()
+        ),
+        "Generated errmap differ from the one checked in"
+    );
+}
