+++ language/tools/move-cli/Cargo.toml
@@ -18,11 +18,9 @@ structopt = "0.3.21"
 
 bcs = "0.1.2"
 bytecode-verifier = { path = "../../bytecode-verifier" }
-diem-vm = { path = "../../diem-vm" }
-diem-framework-releases = { path = "../../diem-framework/releases" }
+
 disassembler = { path = "../disassembler" }
 diem-workspace-hack = { path = "../../../common/workspace-hack" }
-diem-types = { path = "../../../types" }
 move-bytecode-utils = { path = "../move-bytecode-utils" }
 move-coverage = { path = "../move-coverage" }
 move-core-types = { path = "../../move-core/types" }
@@ -33,7 +31,6 @@ move-vm-runtime = { path = "../../move-vm/runtime" }
 read-write-set = { path = "../read-write-set" }
 resource-viewer = { path = "../resource-viewer" }
 move-binary-format = { path = "../../move-binary-format" }
-vm-genesis = { path = "../vm-genesis" }
 walkdir = "2.3.1"
 
 [dev-dependencies]
